[#46] feat : 마이페이지 포트폴리오 수정 작업 완료
